Move BrowserType to models
